Uploading project 3 files on data visualisation.
Includes .ipynb file of data exploration containing univariate, bivariate and multivariate explorations of the prosper loan dataset. The second file is an HTML slide deck which details the explanatory analysis and lastly, there's a README.md file detailing the contents of the dataset and the observations made during the analyses.